typos and question
I don't see how I can open an Issue here. 
I was wondering if it will be possible to test the models that you described in the paper, your experiments and the performance comparison sound extremely compelling. It would be great if it could be possible to test and verify. Especially for MVCNN and RotationNet.